[REFACTOR] Move user keys to their own module
