+++ components/core/src/crypto/keys/cache.rs
@@ -2,13 +2,13 @@ use super::ring_key::RingKey;
 use crate::{crypto::{hash,
                      keys::{box_key_pair::{OriginPublicEncryptionKey,
                                            ServicePublicEncryptionKey,
-                                           ServiceSecretEncryptionKey,
-                                           UserPublicEncryptionKey,
-                                           UserSecretEncryptionKey},
+                                           ServiceSecretEncryptionKey},
                             sig_key_pair::{PublicOriginSigningKey,
                                            SecretOriginSigningKey},
                             Key,
-                            NamedRevision}},
+                            NamedRevision,
+                            UserPublicEncryptionKey,
+                            UserSecretEncryptionKey}},
             error::{Error,
                     Result},
             fs::AtomicWriter};
